Feature/resultados parciais paredao

+++ repositories/elimination-repository.js
@@ -2,6 +2,7 @@
 
 var mongoose = require('mongoose');
 var eliminationModel = require('../models/elimination-model').eliminationModel;
+var objectId =  mongoose.Types.ObjectId;
 
 module.exports = new class EliminationRepository {
 
@@ -10,6 +11,11 @@ module.exports = new class EliminationRepository {
     };
 
     delete(id) {
+
+        if (! objectId.isValid(id)) {
+            throw Error('This id is invalid.');
+        }
+
         return eliminationModel.findByIdAndRemove(id);
     }
 
@@ -18,13 +24,22 @@ module.exports = new class EliminationRepository {
     };
 
     getById(id) {
+
+        if (! objectId.isValid(id)) {
+            throw Error('This id is invalid.');
+        }
+
         return eliminationModel.findById(id);
     };
 
-    getResume(data) {
+    getResume(id) {
+
+        if (! objectId.isValid(id)) {
+            throw Error('This id is invalid.');
+        }
 
         return eliminationModel.aggregate([
-            { $match: {"_id": mongoose.Types.ObjectId(data.id) } },
+            { $match: {"_id": objectId(id) } },
             { $unwind: '$participants' },
             { $unwind: '$participants.votes' },
             { $project: {_id: '$name', participants: '$participants'} },
@@ -39,12 +54,76 @@ module.exports = new class EliminationRepository {
 
     }
 
+    getResumeByHour(id) {
+
+        if (! objectId.isValid(id)) {
+            throw Error('This id is invalid.');
+        }
+
+        return eliminationModel.aggregate([
+            { $match: {"_id": objectId(id) } },
+            { $unwind: '$participants' },
+            { $unwind: '$participants.votes' },
+            {
+                $group: {
+                    _id: {
+                        d: { $dayOfMonth: "$participants.votes.createdAt" },
+                        m: { $month: "$participants.votes.createdAt" },
+                        h: { $hour: "$participants.votes.createdAt" },
+                    },
+                    total: { $sum: 1 }
+                }
+            },
+            { $sort:{ "_id.m": 1 , "_id.d": 1, "_id.h": 1} }
+        ]);
+
+
+    }
+
+    getResumeByParticipant(data) {
+
+        if (! objectId.isValid(data.id) || ! objectId.isValid(data.participantId)) {
+            throw Error('This id is invalid.');
+        }
+
+        return eliminationModel.aggregate([
+            { $unwind: '$participants'},
+            { $unwind: '$participants.votes'},
+            { $match: {"_id": objectId(data.id) }},
+            { $group: {
+                    _id: "$name",
+                    total: {'$sum': 1 },
+                    totalOfParticipant: {
+                        $sum: {
+                            $cond: [
+                                { "$eq": [ "$participants._id", objectId(data.participantId) ] },
+                                1,
+                                0
+                            ]
+                        }
+                    }
+                }}
+        ]);
+
+
+    };
+
     update(data) {
+
+        if (! objectId.isValid(data.id)) {
+            throw Error('This id is invalid.');
+        }
+
         return eliminationModel.findByIdAndUpdate(data.id, {$set: data}, {new: true});
     };
 
-    //TODO Validate if ids are of type ObjectType
     vote(data) {
+
+        // TODO validate if the elimination is open to vote
+
+        if (! objectId.isValid(data.id)) {
+            throw Error('This id is invalid.');
+        }
 
         return eliminationModel.findOneAndUpdate({
                 "_id": data.id,
